requirements.txt
google-cloud-logging==3.11.2
google-cloud-monitoring==2.22.2